refactor(documentation): moved the documentation scaffolder into the new  documentation folder
